requirements.txt
aiofiles==0.3.2
aiohttp==3.0.6
aiopg==0.13.2
async-timeout==2.0.0
attrs==17.4.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.4.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.1
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.10.0
httplib2==0.10.3
httptools==0.0.10
idna==2.6
idna-ssl==1.0.0
ipaddress==1.0.19
kubernetes==5.0.0
multidict==4.1.0
oauth2client==3.0.0
oauthlib==2.0.6
pamqp==1.6.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==5.4.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.0
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2018.3
PyYAML==3.12
rabbitpy==1.0.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
sanic==0.7.0
six==1.11.0
ujson==1.35
urllib3==1.22
uvloop==0.9.1
websocket-client==0.47.0
websockets==4.0.1
yarl==1.1.1
gevent==1.2.2